Remove lance_ prefixes, remove unused code
