+++ unstructured_ingest/v2/processes/connectors/lancedb/__init__.py
@@ -2,14 +2,14 @@ from __future__ import annotations
 
 from unstructured_ingest.v2.processes.connector_registry import add_destination_entry
 
-from .lancedb_aws import CONNECTOR_TYPE as LANCEDB_S3_CONNECTOR_TYPE
-from .lancedb_aws import lancedb_aws_destination_entry
-from .lancedb_azure import CONNECTOR_TYPE as LANCEDB_AZURE_CONNECTOR_TYPE
-from .lancedb_azure import lancedb_azure_destination_entry
-from .lancedb_gcp import CONNECTOR_TYPE as LANCEDB_GCS_CONNECTOR_TYPE
-from .lancedb_gcp import lancedb_gcp_destination_entry
-from .lancedb_local import CONNECTOR_TYPE as LANCEDB_LOCAL_CONNECTOR_TYPE
-from .lancedb_local import lancedb_local_destination_entry
+from .aws import CONNECTOR_TYPE as LANCEDB_S3_CONNECTOR_TYPE
+from .aws import lancedb_aws_destination_entry
+from .azure import CONNECTOR_TYPE as LANCEDB_AZURE_CONNECTOR_TYPE
+from .azure import lancedb_azure_destination_entry
+from .gcp import CONNECTOR_TYPE as LANCEDB_GCS_CONNECTOR_TYPE
+from .gcp import lancedb_gcp_destination_entry
+from .local import CONNECTOR_TYPE as LANCEDB_LOCAL_CONNECTOR_TYPE
+from .local import lancedb_local_destination_entry
 
 add_destination_entry(LANCEDB_S3_CONNECTOR_TYPE, lancedb_aws_destination_entry)
 add_destination_entry(LANCEDB_AZURE_CONNECTOR_TYPE, lancedb_azure_destination_entry)
