+++ unstructured_ingest/v2/processes/connectors/lancedb/gcp.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from typing import TYPE_CHECKING
 
 from pydantic import Field, Secret
 
@@ -13,31 +12,29 @@ from unstructured_ingest.v2.processes.connectors.lancedb.lancedb import (
     LanceDBUploadStagerConfig,
 )
 
-CONNECTOR_TYPE = "lancedb_s3"
-
-if TYPE_CHECKING:
-    pass
+CONNECTOR_TYPE = "lancedb_gcs"
 
 
-class LanceDBS3AccessConfig(AccessConfig):
-    aws_access_key_id: str = Field(description="The AWS access key ID to use.")
-    aws_secret_access_key: str = Field(description="The AWS secret access key to use.")
+class LanceDBGCSAccessConfig(AccessConfig):
+    google_service_account_key: str = Field(
+        description="The serialized google service account key."
+    )
 
 
-class LanceDBS3ConnectionConfig(LanceDBConnectionConfig):
-    access_config: Secret[LanceDBS3AccessConfig]
+class LanceDBGCSConnectionConfig(LanceDBConnectionConfig):
+    access_config: Secret[LanceDBGCSAccessConfig]
 
 
 @dataclass
-class LanceDBS3Uploader(LanceDBUploader):
+class LanceDBGSPUploader(LanceDBUploader):
     upload_config: LanceDBUploaderConfig
-    connection_config: LanceDBS3ConnectionConfig
+    connection_config: LanceDBGCSConnectionConfig
     connector_type: str = CONNECTOR_TYPE
 
 
-lancedb_aws_destination_entry = DestinationRegistryEntry(
-    connection_config=LanceDBS3ConnectionConfig,
-    uploader=LanceDBS3Uploader,
+lancedb_gcp_destination_entry = DestinationRegistryEntry(
+    connection_config=LanceDBGCSConnectionConfig,
+    uploader=LanceDBGSPUploader,
     uploader_config=LanceDBUploaderConfig,
     upload_stager_config=LanceDBUploadStagerConfig,
     upload_stager=LanceDBUploadStager,
